second commit

+++ src/main/java/guru/springframework/corporate/dto/response/ServicesResponse.java
@@ -1,5 +1,6 @@
 package guru.springframework.corporate.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ServicesResponse {
@@ -8,7 +9,7 @@ public class ServicesResponse {
     private String title;
     private String description;
     private String iconUrl;
-    private Double price;
+    private BigDecimal price;
     private boolean isPublished;
     private LocalDateTime createdAt;
 
@@ -25,8 +26,8 @@ public class ServicesResponse {
     public String getIconUrl() { return iconUrl; }
     public void setIconUrl(String iconUrl) { this.iconUrl = iconUrl; }
 
-    public Double getPrice() { return price; }
-    public void setPrice(Double price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
 
     public boolean isPublished() { return isPublished; }
     public void setPublished(boolean published) { isPublished = published; }
